added tauCeti asteroids debris correct position
